feat: refactor color handling by moving color definitions to a separate module
