+++ src/lib/colors.js
@@ -1,0 +1,12 @@
+import chalk from 'chalk'
+
+export const COLORS = {
+  note: chalk.gray.italic,
+  json: chalk.gray,
+  info: chalk.yellow,
+  success: chalk.green,
+  hint: chalk.blue,
+  error: chalk.red,
+  debug: chalk.gray,
+  message: chalk.gray
+}
